Add more wrappers

+++ 1.12/me/deftware/client/framework/Wrappers/Render/IFontRenderer.java
@@ -26,12 +26,12 @@ public class IFontRenderer {
 		return Minecraft.getMinecraft().fontRendererObj.getStringWidth(string);
 	}
 	
-	public static void drawStringWithShadow(String text, int x, int y, IFontRendererObject font) {
-		font.drawStringWithShadow(text, x, y, 0xFFFFFF);
+	public static void drawStringWithShadow(String text, float x, float y, IFontRendererObject font) {
+		font.drawStringWithShadow(text, (int) x, (int) y, 0xFFFFFF);
 	}
 	
-	public static void drawStringWithShadow(String text, int x, int y, int color, IFontRendererObject font) {
-		font.drawStringWithShadow(text, x, y, color);
+	public static void drawStringWithShadow(String text, float x, float y, int color, IFontRendererObject font) {
+		font.drawStringWithShadow(text, (int) x, (int) y, color);
 	}
 	
 	public static int getFontHeight(IFontRendererObject font) {
